+++ src/snlsgn.c
@@ -1,7 +1,7 @@
 #include "config.h"
 #include "sylvan.h"
 
-long snlsgn(long x)
+int snlsgn(long x)
 {
 	return (x > 0) - (x < 0);
 }
