+++ src/snqsgn.c
@@ -1,7 +1,7 @@
 #include "config.h"
 #include "sylvan.h"
 
-unsigned long snqsgn(unsigned long x)
+int snqsgn(unsigned long x)
 {
 	return x > 0;
 }
