+++ src/snzsgn.c
@@ -2,7 +2,7 @@
 #include <stddef.h>
 #include "sylvan.h"
 
-size_t snzsgn(size_t x)
+int snzsgn(size_t x)
 {
 	return x > 0;
 }
